supabase/functions/python-blueprint-engine/requirements.txt

openai>=1.14.0
hdkit>=0.3.0
pyswisseph
requests
python-dotenv
